added unified hybrid queries (overlap with #287)
