+++ CONTRIBUTORS.md
@@ -88,7 +88,7 @@ We aim to respond within a few days. Be patient with us, and we'll be patient wi
 Thank you to everyone who has contributed to Grafeo!
 
 - **CorvusYe** ([@CorvusYe](https://github.com/CorvusYe)): Dart bindings ([#138](https://github.com/GrafeoDB/grafeo/pull/138)), single-file `.grafeo` format feature request ([#139](https://github.com/GrafeoDB/grafeo/issues/139))
-- **temporaryfix** ([@temporaryfix](https://github.com/temporaryfix)): CompactStore columnar read-optimized store, RFC ([#199](https://github.com/GrafeoDB/grafeo/issues/199)), implementation ([#204](https://github.com/GrafeoDB/grafeo/pull/204)); BM25 text scan operator, hybrid query planner pushdown, Float64/Float32Vector column codecs, post-compact index fix ([#286](https://github.com/GrafeoDB/grafeo/pull/286))
+- **temporaryfix** ([@temporaryfix](https://github.com/temporaryfix)): CompactStore columnar read-optimized store, RFC ([#199](https://github.com/GrafeoDB/grafeo/issues/199)), implementation ([#204](https://github.com/GrafeoDB/grafeo/pull/204)); Float32Vector codec and post-compact index fix ([#286](https://github.com/GrafeoDB/grafeo/pull/286)); unified hybrid queries proposa, BM25 `TextScanOperator`, planner pushdown for vector + text predicates, top-K rewrite, cost model ([#287](https://github.com/GrafeoDB/grafeo/pull/287))
 - **Imaclean74** ([@Imaclean74](https://github.com/Imaclean74)): CompactStore multi-label-pair edge type fix ([#225](https://github.com/GrafeoDB/grafeo/pull/225))
 - **Michaelzag** ([@Michaelzag](https://github.com/Michaelzag)): `cypher` feature `gql` dependency fix ([#233](https://github.com/GrafeoDB/grafeo/pull/233)), schema type extraction proposal ([#234](https://github.com/GrafeoDB/grafeo/issues/234)), Python named graph management ([#243](https://github.com/GrafeoDB/grafeo/pull/243)), Python per-transaction CDC ([#244](https://github.com/GrafeoDB/grafeo/pull/244)), graph/schema context validation ([#246](https://github.com/GrafeoDB/grafeo/pull/246)). **grafeo-server**: backup/restore endpoints with ArcSwap database handle ([server#53](https://github.com/GrafeoDB/grafeo-server/pull/53)), sync pull CDC error handling and test fixes ([server#55](https://github.com/GrafeoDB/grafeo-server/pull/55)), backup chain API adoption and database file rename ([server#60](https://github.com/GrafeoDB/grafeo-server/pull/60)), admin UI redesign with backup labels, studio auth, restore-to-epoch, cross-database restore, and graph view fix ([server#63](https://github.com/GrafeoDB/grafeo-server/pull/63))
