+++ crates/grafeo-core/Cargo.toml
@@ -90,3 +90,4 @@ workspace = true
 [[bench]]
 name = "index_bench"
 harness = false
+required-features = ["vector-index"]
