+++ crates/grafeo-engine/src/database/section_consumer.rs
@@ -65,6 +65,7 @@ use std::path::PathBuf;
 /// without a full checkpoint + mmap cycle. The [`can_spill`](MemoryConsumer::can_spill)
 /// method returns `true` for mmap-able index sections, signaling that future
 /// tiered storage support will enable actual spilling.
+#[cfg(feature = "lpg")]
 pub struct SectionConsumer {
     name: String,
     section: Arc<dyn Section>,
@@ -73,6 +74,7 @@ pub struct SectionConsumer {
     mmap_able: bool,
 }
 
+#[cfg(feature = "lpg")]
 impl SectionConsumer {
     /// Creates a consumer for the given section.
     ///
@@ -102,6 +104,7 @@ impl SectionConsumer {
     }
 }
 
+#[cfg(feature = "lpg")]
 impl MemoryConsumer for SectionConsumer {
     fn name(&self) -> &str {
         &self.name
